+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# [NEXT]
+- Adds `animateZoom` method in `Camera`.
+
 # [2.2.5]
 - Fix crash mentioned in th issue [#255](https://github.com/RafaelBarbosatec/bonfire/issues/225)
 - Fix crash mentioned in th issue [#227](https://github.com/RafaelBarbosatec/bonfire/issues/227)