+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# [NEXT]
+- Experimental State manager. Example [here](https://github.com/RafaelBarbosatec/bonfire/issues/218#issuecomment-1058121200)
+    - abstract class`StateController`;
+    - mixin `UseStateController`;
+    - Widget `StateControllerConsumer`;
+    - `BonfireInjector`;
+
+
 # [2.2.5]
 - Fix crash mentioned in th issue [#255](https://github.com/RafaelBarbosatec/bonfire/issues/225)
 - Fix crash mentioned in th issue [#227](https://github.com/RafaelBarbosatec/bonfire/issues/227)